requirements.txt
# Яндекс Спич — Зависимости проекта
# Установка: pip install -r requirements.txt

# === Обязательные ===
requests>=2.28.0
python-docx>=0.8.11
pydub>=0.25.1
pymorphy2>=0.9.1
rapidfuzz>=3.0.0
boto3>=1.26.0

# === Веб-интерфейс ===
flask>=2.3.0

# === Разработка и тестирование ===
pytest>=6.0
pytest-cov>=4.0

# === Опциональные ===
# tqdm>=4.60.0  # Прогресс-бары (необязательно)
